switch to laptop and ubuntu 22.04
